[Refactor] TourAPI 설정 값 조회 방식 변경
이제 상수값 조회가 아닌,
`application-openapi.yml` 에서 가져와 사용합니다
