+++ app/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 spring:
   profiles:
     include:
-      common
+      common, openapi
     active: ${APP_PROFILE}
